view_created & classes 05: hotfixes, rename Vacancies to VacanciesCommands, create Vacancies module, replace Vacancy class, pochinil importi, fix methods
