+++ handlers/commands.py
@@ -1,4 +1,4 @@
-from aiogram import Router, Bot
+from aiogram import Router
 from aiogram.types import Message
 from aiogram.filters import Command
 from aiogram.fsm.context import FSMContext
@@ -6,7 +6,7 @@ from aiogram.fsm.context import FSMContext
 from fsm.statesform import StapesForm as sf
 
 from keyboards.inline_keyboards import *
-from classes import User, user_commands, vac_commands
+from classes import User, Vacancy, vac_commands, db_commands
 import asyncio
 from assets import texts
 __all__ = ["command_start", "command_choice", "command_create_vacancy"]
@@ -15,7 +15,7 @@ router = Router()
 
 
 @router.message(Command(commands=['start']))
-async def command_start(message: Message, bot: Bot, state: FSMContext):
+async def command_start(message: Message, state: FSMContext):
     if await state.get_state() is None:
         await message.reply(texts.welcome_text)
         await asyncio.sleep(0.3)
@@ -25,7 +25,7 @@ async def command_start(message: Message, bot: Bot, state: FSMContext):
                     username=message.from_user.username,
                     fullname=message.from_user.full_name)
 
-        await user.add_to_db()
+        await db_commands.add_user_to_db(user)
     else:
         await message.answer(text=texts.default_state_warn)
 
@@ -57,14 +57,14 @@ async def command_show_favorites(message: Message, state: FSMContext):
         user_liked_vacancies = await vac_commands.get_user_likes(user)
 
         if user_liked_vacancies:
-            for vacancy in user_liked_vacancies:
+            for vacancy_values in user_liked_vacancies:
+                vacancy = Vacancy(values=await db_commands.row_to_dict(vacancy_values))
 
                 text = await vac_commands.to_text(vacancy=vacancy,
                                                   type_descr="short")
-                liked_vacancy_id = vacancy[0]
 
                 await message.answer(text=text,
-                                     reply_markup=await create_inkb(id=liked_vacancy_id,
+                                     reply_markup=await create_inkb(id=vacancy.id,
                                                                     is_next=False,
                                                                     btn_like_nlike="nlike",
                                                                     btn_more_less="more"))
@@ -84,12 +84,14 @@ async def command_show_created_vacancies(message: Message, state: FSMContext):
         created_user_vacancies = await vac_commands.get_user_creates(user)
 
         if created_user_vacancies:
-            for vacancy in created_user_vacancies:
-                text = await vac_commands.to_text(vacancy=vacancy, type_descr="short")
+            for vacancy_values in created_user_vacancies:
+                vacancy = Vacancy(values=await db_commands.row_to_dict(vacancy_values))
 
-                created_vacancy_id = vacancy[0]
+                text = await vac_commands.to_text(vacancy=vacancy,
+                                                  type_descr="short")
+
                 await message.answer(text=text,
-                                     reply_markup=await create_inkb_for_employer(id=created_vacancy_id,
+                                     reply_markup=await create_inkb_for_employer(id=vacancy.id,
                                                                                  btn_more_less="more"))
         else:
             await message.answer(texts.no_created)
